refactor(shell/sidebar): extract workspaces panel + tab-context-menu (PR 12/12)
Moves ~670 LOC out of shell/js/sidebar/index.js into two new ES modules,
finishing the sidebar split and leaving index.js as pure orchestration.

- panels/workspaces.js: workspace list/form/icon grid, tab-bar filtering,
  openWorkspacePanel. Sidebar re-render is wired via setWorkspacesRender()
  from init() to avoid a circular import with index.js.
- tab-context-menu.js (sibling to panels/ — it's a floating menu, not a
  panel): the full tab right-click menu and its quick-link helpers.
  window.__tandemShowTabContextMenu is installed as a module side effect.

Pure structural move — no behavior change. activateItem now routes the
WORKSPACE_PANEL_ID click path explicitly, mirroring the pinboard pattern.

sidebar/index.js: 956 → 332 LOC.
